requirements.txt
streamlit
feedparser
google-generativeai
PyGithub


